V0.0.9
Login

+++ admin/Noticias/apagar.php
@@ -1,5 +1,5 @@
 <?php
-//if(isset($_SESSION['adm'])){ 
+if(isset($_SESSION['adm'])){ 
     include "conecta.php";
     if(is_numeric($_GET["id"])){
         $SQL = "DELETE FROM noticias WHERE id = ".$_GET["id"];
@@ -10,5 +10,9 @@
         }
     }
 	mysql_close();
-//}
+}else{
+   include("header.php");
+   echo'<div class="alert alert-danger" role="alert">Você não tem acesso a essa página!</div>';
+   include('footer.php');
+}
 ?>
